+++ goopylib/src/Platform/OpenGL/VertexArray.cpp
@@ -5,11 +5,15 @@ namespace gp {
             : BaseVertexArray() {
         glGenVertexArrays(1, &m_RendererID);
         glBindVertexArray(m_RendererID);
+
+        GP_CORE_TRACE("Initialising Vertex Array {0}", m_RendererID);
     }
 
     VertexArray::VertexArray(uint32_t *indices, int count) : BaseVertexArray() {
         glGenVertexArrays(1, &m_RendererID);
         glBindVertexArray(m_RendererID);
+
+        GP_CORE_TRACE("Initialising Vertex Array {0}", m_RendererID);
 
         setIndexBuffer(std::make_shared<IndexBuffer>(indices, count));
     }
@@ -18,10 +22,13 @@ namespace gp {
         glGenVertexArrays(1, &m_RendererID);
         glBindVertexArray(m_RendererID);
 
+        GP_CORE_TRACE("Initialising Vertex Array {0}", m_RendererID);
+
         setIndexBuffer(std::make_shared<IndexBuffer>(indices));
     }
 
     VertexArray::~VertexArray() {
+        GP_CORE_TRACE("Deallocating Vertex Array {0}", m_RendererID);
         glDeleteVertexArrays(1, &m_RendererID);
     }
 
@@ -30,6 +37,7 @@ namespace gp {
     }
 
     void VertexArray::unbind() const {
+        GP_CORE_WARN("Unbinding Vertex Arrays");
         glBindVertexArray(0);
     }
 
@@ -42,9 +50,9 @@ namespace gp {
         bind();
         vertexBuffer.bind();
 
-        const auto &layout = vertexBuffer.getLayout();
+        const auto layout = vertexBuffer.getLayout();
 
-        for (const auto &element: layout) {
+        for (const auto element: layout) {
             ShaderDataType type = element.getType();
 
             switch (type) {
@@ -52,6 +60,9 @@ namespace gp {
                 case ShaderDataType::Float2:
                 case ShaderDataType::Float3:
                 case ShaderDataType::Float4: {
+
+                    GP_CORE_TRACE("Vertex Array Attribute index={0}, size={1}, type=GL_FLOAT, normalized={2}, stride={3}, offset={4}",
+                                  m_VertexBufferIndex, element.getCount(), element.isNormalized(), layout.getStride(), (const void *) element.m_Offset);
 
                     glEnableVertexAttribArray(m_VertexBufferIndex);
                     glVertexAttribPointer(m_VertexBufferIndex,
@@ -68,10 +79,14 @@ namespace gp {
                 case ShaderDataType::Int3:
                 case ShaderDataType::Int4:
                 case ShaderDataType::Bool: {
+
+                    GP_CORE_TRACE("Vertex Array Attribute index={0}, size={1}, type={2}, stride={3}, offset={4}",
+                                  m_VertexBufferIndex, element.getCount(), shaderOpenGLType(type), layout.getStride(), (const void *) element.m_Offset);
+
                     glEnableVertexAttribArray(m_VertexBufferIndex);
                     glVertexAttribIPointer(m_VertexBufferIndex,
                                            element.getCount(),
-                                           shaderOpenGLType(element.getType()),
+                                           shaderOpenGLType(type),
                                            layout.getStride(),
                                            (const void *) element.m_Offset);
                     m_VertexBufferIndex++;
@@ -86,6 +101,10 @@ namespace gp {
                     int stride = layout.getStride();
 
                     for (int i = 0; i < count; i++) {
+
+                        GP_CORE_TRACE("Vertex Array Attribute index={0}, size={1}, type={2}, normalized={3}, stride={4}, offset={5}",
+                                      m_VertexBufferIndex, element.getCount(), shaderOpenGLType(type), normalized, layout.getStride(), (const void *) (element.m_Offset + sizeof(float) * count * i));
+
                         glEnableVertexAttribArray(m_VertexBufferIndex);
                         glVertexAttribPointer(m_VertexBufferIndex,
                                               count, glType, normalized, stride,