+++ goopylib/src/extension/renderer/shader.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "extension/util.h"
+
+#include "goopylib/Core/Shader.h"