+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Added log mesages to the Buffer classes
 * Added a Vertex Array draw method
 * Vertex Arrays can now create their own Index Buffers
+* Added Python wrappers for VertexArray, VertexBuffer, IndexBuffer, and Shader
 
 #### 9 December, 2022 - 14290 lines of code